discord/requirements.txt
discord.py
python-dotenv
requests
pycryptodome